Add a User-Agent header to ParseURL()
I changed the .Get() call to .Do() here for a header insertion

This is a cheap edit, but I've discovered that on GETs the User-Agent is being sent blank.  This is a problem for RSS feeds from Reddit, as Reddit will send back phony 429s  (Rate Limit) responses because the User-Agent is not "valid".   Might want to expose this further if people need to imitate a User-Agent to get around a server-side restriction.