Support marshal to xml for atom
This adds the necessary tags to the atom structs so that someone can
call xml.Marshal on an atom feed to generate the appropriate xml.

Note that Go doesn't support self-closing tags, which is why you see
that the link tag in the testdata is using an empty tag instead.

I am not marshaling some fields that are calculated by the parser:
Updated/PublishedParsed, Extensions, and Version. xml:lang on the root
feed node isn't being marhalled either because it doesn't fit well with
how the data is being represented. This seems like a good first step
towards unobtrusively adding marshalling without impacting the core of
gofeed.